requirements.txt
nltk
pyyaml
